// src/types.ts
export type Person = {
    id: string;
    name: string;
    role: string;
    description: string;
    image: string;
  };
  
export type Assignment = {
    id: string;
    name: string;
    role: string;
    description: string;
    image: string;
 };
  
 export type Notification = {
    id: string;
    name: string;
    role: string;
    description: string;
    image: string;
 };